uhakikiai_api/requirements.txt
fastapi
uvicorn
supabase
python-multipart
passlib[bcrypt]
easyocr
opencv-python-headless
numpy
python-dotenv
pillow